Move behavior classes into new subfolder of engine.
