--- src/engine/ai/behavior-wander.ts
@@ -1,1 +1,0 @@
-export const wanderBetweenRooms = 5
