+++ src/engine/world.ts
@@ -203,8 +203,6 @@ export class World extends TypedEventEmitter<WorldEvents> {
     }
 
     if (creature !== undefined && action !== undefined) {
-      // eslint-disable-next-line no-console
-      console.log('act', action)
       // TODO: examine success/failure return value
       const result = action.execute(this)
